+++ fiddles/angular2-cli/fiddle-0000-Template/e2e/app.e2e.ts
@@ -1,0 +1,14 @@
+import { Fiddle0000TemplatePage } from './app.po';
+
+describe('fiddle-0000-template App', function() {
+  let page: Fiddle0000TemplatePage;
+
+  beforeEach(() => {
+    page = new Fiddle0000TemplatePage();
+  })
+
+  it('should display message saying app works', () => {
+    page.navigateTo();
+    expect(page.getParagraphText()).toEqual('fiddle-0000-template works!');
+  });
+});