+++ fiddles/angular2-cli/fiddle-0000-Template/e2e/app.po.ts
@@ -1,0 +1,9 @@
+export class Fiddle0000TemplatePage {
+  navigateTo() {
+    return browser.get('/');
+  }
+
+  getParagraphText() {
+    return element(by.css('fiddle-0000-template-app h1')).getText();
+  }
+}