+++ fiddles/angular2-cli/fiddle-0000-Template/src/app/index.ts
@@ -1,0 +1,2 @@
+export {environment} from './environment';
+export {Fiddle0000TemplateAppComponent} from './fiddle-0000-template.component';